+++ src/app/onboarding/_components/SignInForm.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { useForm } from "react-hook-form";
+import { useForm, FieldValues } from "react-hook-form";
 import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Button } from "../common/button";
-import { Input } from "../common/input";
 import { useToast } from "@/utils/useToast";
-import { handleInvalidForm, handleSignInSubmit } from "@/utils/signInHandle";
+import { Button } from "@/components/common/button";
+import { Input } from "@/components/common/input";
+import { UseFormSetError } from "react-hook-form";
 
 const schema = z.object({
   email: z.string().email("아이디는 이메일 주소 형태로 입력해주세요."),
@@ -29,6 +29,38 @@ const SignInForm = () => {
   const password = watch("password") ?? "";
   const isInputFilled = email.trim() !== "" && password.trim() !== "";
 
+  const handleSignInSubmit = async (data: FieldValues, setError: UseFormSetError<any>) => {
+    try {
+      const userExists = false;
+      const passwordCorrect = false;
+
+      if (!userExists) {
+        const message = "가입되지 않은 이메일 주소예요.";
+        setError("email", { type: "manual", message });
+        toast({ variant: "error", description: message });
+        return;
+      }
+
+      if (!passwordCorrect) {
+        const message = "아이디 또는 비밀번호가 맞지 않아요.\n다시 확인해주세요.";
+        setError("password", { type: "manual", message });
+        toast({ variant: "error", description: message });
+        return;
+      }
+
+      console.log("로그인 성공", data);
+    } catch (err) {
+      console.error("로그인 실패", err);
+    }
+  };
+
+  const handleInvalidForm = (errors: Record<string, any>) => {
+    const firstError = errors.email?.message || errors.password?.message;
+    if (firstError) {
+      toast({ variant: "error", description: firstError });
+    }
+  };
+
   return (
     <section className="w-full flex flex-col gap-3">
       <div className="w-full px-4 flex flex-col gap-2">
@@ -38,8 +70,8 @@ const SignInForm = () => {
 
       <form
         onSubmit={handleSubmit(
-          data => handleSignInSubmit(data, setError, toast),
-          errors => handleInvalidForm(errors, toast)
+          data => handleSignInSubmit(data, setError),
+          errors => handleInvalidForm(errors)
         )}
         className="flex flex-col gap-5 mt-2">
         <div className="flex flex-col gap-2">
