+++ src/app/page.tsx
@@ -1,6 +1,5 @@
 import { Container } from "@/components";
 import Link from "next/link";
-import OnBoardingFooter from "@/components/onBoarding/OnBoardingFooter";
 
 export default function Home() {
   return (
@@ -10,7 +9,6 @@ export default function Home() {
           <h1 className="text-h1">Zaply</h1>
           <p className="text-b1R">Create once. Spread with Zaply</p>
           <Link href="/policy">회원가입</Link>
-          <OnBoardingFooter />
         </div>
       </div>
     </Container>
